+++ project/history.go
@@ -29,7 +29,22 @@ func FullVersion() string {
 // To cut a new release add a release to the front of this slice then run the
 // release tagging script: ./scripts/tag_release.sh
 var History relic.ImmutableHistory = relic.NewHistory("Monax Hoard", "https://github.com/monax/hoard").
-	MustDeclareReleases(
+	MustDeclareReleases("",
+		``,
+		"3.0.0 - 2019-02-26",
+		`### Changed
+- [PROTO] Renamed symmetric grant SecretID to PublicID
+- [PROTO] Renamed openpgp grant ID to PrivateID
+
+### Fixed
+- [GRANTS] Throw an exception if symmetric secret for ID cannot be found
+
+### Added
+- [NODEJS] Added integration tests including test for symmetric secrets
+- [GRANTS] Added openpgp grants example
+- [CLI] Added ability to configure secrets on command line with hoard config <config> --secret
+- 
+`,
 		"2.0.0 - 2019-02-21",
 		`### Changed
 - [PROTO] Upper case field names in protobuf