+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 # [Monax Hoard](https://github.com/monax/hoard) Changelog
+## [Unreleased]
+
+
+## [3.0.0] - 2019-02-26
+### Changed
+- [PROTO] Renamed symmetric grant SecretID to PublicID
+- [PROTO] Renamed openpgp grant ID to PrivateID
+
+### Fixed
+- [GRANTS] Throw an exception if symmetric secret for ID cannot be found
+
+### Added
+- [NODEJS] Added integration tests including test for symmetric secrets
+- [GRANTS] Added openpgp grants example
+- [CLI] Added ability to configure secrets on command line with hoard config <config> --secret
+- 
+
+
 ## [2.0.0] - 2019-02-21
 ### Changed
 - [PROTO] Upper case field names in protobuf
@@ -67,6 +85,8 @@ This is the first Hoard open source release and includes:
 	- Hoar-Daemon hoard
 	- Hoar-Control hoarctl CLI
 
+[Unreleased]: https://github.com/monax/hoard/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/monax/hoard/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/monax/hoard/compare/v1.1.5...v2.0.0
 [1.1.5]: https://github.com/monax/hoard/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/monax/hoard/compare/v1.1.3...v1.1.4